+++ src/main/java/frc/robot/subsystems/AprilTagOdometry.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.odometry;
+package frc.robot.subsystems;
 
 import java.io.IOException;
 import java.util.List;
@@ -24,7 +24,6 @@ import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
-import frc.robot.subsystems.SwerveDrive;
 
 /**
  * This class will manage processes related to photon vision,
